Add verification methods all status code verification form addclient

+++ src/Controller/ClientsController.php
@@ -28,6 +28,8 @@ class ClientsController extends AbstractController
   public function ShowClients($usersid, SerializerInterface $serialize, Request $request): Response
   {
 
+    if ($request->isMethod('GET')) {
+
       $repository = $this->getDoctrine()->getRepository(Clients::class);
       $clients = $repository->findAllClients($usersid);
 
@@ -37,51 +39,128 @@ class ClientsController extends AbstractController
       $response->headers->set('Content-Type', 'application/json');
 
       return $response;
+    }
 
+    else {
+      return api_response('Utiliser la methode GET', 405);
+    }
   }
 
   public function ShowSingleClients($usersid, $clientsid, SerializerInterface $serialize, Request $request): Response
   {
 
+    if ($request->isMethod('GET')) {
+
       $repository = $this->getDoctrine()->getRepository(Clients::class);
       $clients = $repository->findSingleClients($usersid, $clientsid);
 
+      if (!empty($clients)) {
       $data = $serialize->serialize($clients, 'json');
 
       $response = new Response($data);
       $response->headers->set('Content-Type', 'application/json');
 
       return $response;
+    }
+      else {
+        return api_response('Ce client n existe pas', 404);
+      }
+    }
 
+    else {
+      return api_response('Utiliser la methode GET', 405);
+    }
   }
 
   public function AddClient($usersid, Request $request): Response
     {
+
+      if ($request->isMethod('POST')) {
+
         $client = new Clients();
 
+        // vérification nom
         $name = htmlspecialchars(filter_input(INPUT_POST, 'name'));
-        $client->setName($name);
+        if (isset($name) && ctype_alpha($name)) {
+          $client->setName($name);
+        }
+        else {
+          return api_response('Nom du client invalide', 400);
+        }
+        // ***
 
+        // vérification prénom
         $firstname = htmlspecialchars(filter_input(INPUT_POST, 'firstname'));
-        $client->setFirstname($firstname);
+        if (isset($firstname) && ctype_alpha($firstname)) {
+          $client->setFirstname($firstname);
+        }
+        else {
+          return api_response('Prénom du client invalide', 400);
+        }
+        // ***
 
+        // vérification email
         $email = htmlspecialchars(filter_input(INPUT_POST, 'email'));
-        $client->setEmail($email);
+        if (isset($email) && filter_var($email, FILTER_VALIDATE_EMAIL)) {
+          $client->setEmail($email);
+        }
+        else {
+          return api_response('Email du client invalide', 400);
+        }
+        // ***
 
+        // vérification numero de rue
         $number = htmlspecialchars(filter_input(INPUT_POST, 'number'));
-        $client->setNumber($number);
+        if (isset($number) && is_numeric($number)) {
+          $client->setNumber($number);
+        }
+        else {
+          return api_response('Numero de rue invalide, utiliser seulement des chiffres', 400);
+        }
+        // ***
 
+        // vérification rue
         $street = htmlspecialchars(filter_input(INPUT_POST, 'street'));
-        $client->setStreet($street);
+        if (isset($street) && ctype_alpha($street)) {
+          $client->setStreet($street);
+        }
+        else {
+          return api_response('Rue du client invalide', 400);
+        }
+        // ***
 
+        // vérification code postal
         $postalCode = htmlspecialchars(filter_input(INPUT_POST, 'postalCode'));
-        $client->setPostalCode($postalCode);
+        if (isset($postalCode) && is_numeric($postalCode)) {
+          $client->setPostalCode($postalCode);
+        }
+        else {
+          return api_response('Code postal invalide', 400);
+        }
+        // ***
 
+        // vérification ville
         $city = htmlspecialchars(filter_input(INPUT_POST, 'city'));
-        $client->setCity($city);
+        if (isset($city) && ctype_alpha($city)) {
+          $client->setCity($city);
+        }
+        else {
+          return api_response('Ville du client invalide', 400);
+        }
+        // ***
 
+        // vérification téléphone
         $tel = htmlspecialchars(filter_input(INPUT_POST, 'tel'));
-        $client->setTel($tel);
+        if (isset($tel) && preg_match("#^0[1-68]([-. ]?[0-9]{2}){4}$#", $tel))
+        {
+        	$meta_carac = array("-", ".", " ");
+        	$tel = str_replace($meta_carac, "", $tel);
+        	$client->setTel($tel);
+        }
+        else {
+          return api_response('Numéro de téléphone du client invalide', 400);
+        }
+        // ***
 
         $repository = $this->getDoctrine()->getRepository(Users::class);
         $user = $repository->findOneBy(['id' => $usersid]);
@@ -90,20 +169,36 @@ class ClientsController extends AbstractController
         $this->em->persist($client);
         $this->em->flush();
 
-        return api_response('OK', 200);
+        return api_response('Le client à été ajouter', 200);
       }
+
+      else {
+        return api_response('Utiliser la methode POST', 405);
+      }
+    }
 
       public function DeleteClient($usersid, $clientsid, Request $request): Response
       {
 
+        if ($request->isMethod('DELETE')) {
+
           $repository = $this->getDoctrine()->getRepository(Clients::class);
           $client = $repository->findOneBy(['id' => $clientsid, 'user' => $usersid]);
 
+          if (!empty($client)) {
           $this->em->remove($client);
           $this->em->flush();
 
-          return api_response('OK', 200);
+          return api_response('Le client à été supprimer', 200);
+          }
 
+          else {
+            return api_response('Ce client n existe pas', 404);
+          }
+        }
+
+        else {
+          return api_response('Utiliser la methode DELETE', 405);
+        }
       }
-
 }
